veritcal splash screen with loading bar
