+++ flutter_application_1/lib/main.dart
@@ -1,7 +1,7 @@
 // ignore_for_file: camel_case_types
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
+//import 'package:flutter/services.dart';
 import 'package:flutter_application_1/pages/splash_screen.dart';
 
 void main() async {
@@ -13,10 +13,10 @@ void main() async {
           messagingSenderId: 'sendid',
           projectId: "kards-c7448",)
           );
-  SystemChrome.setPreferredOrientations([
-    DeviceOrientation.landscapeLeft,
-    DeviceOrientation.landscapeRight,
-  ]);
+  // SystemChrome.setPreferredOrientations([
+  //   DeviceOrientation.portraitUp,
+  //   DeviceOrientation.portraitDown,
+  // ]);
   runApp(const myApp());
 }
