--- flutter_application_1/lib/pages/Host.dart
@@ -1,19 +1,0 @@
-import 'package:flutter/material.dart';
-
-class Host extends StatelessWidget {
-  const Host({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return const Scaffold(
-      body : Text(
-        "Single Player Screen \n I hate my life",
-        style: TextStyle(
-          color: Colors.black, 
-          fontSize: 50,
-          fontFamily: 'RedRose',
-        ),
-      ),
-    );
-  }
-}
