--- flutter_application_1/lib/pages/Join.dart
@@ -1,19 +1,0 @@
-import 'package:flutter/material.dart';
-
-class Join extends StatelessWidget {
-  const Join({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return const Scaffold(
-      body : Text(
-        "Single Player Screen \n Built with both the air and constitution of a Victorian Noblity",
-        style: TextStyle(
-          color: Colors.black, 
-          fontSize: 50,
-          fontFamily: 'RedRose',
-        ),
-      ),
-    );
-  }
-}
